+++ mangadex_library/lib/models/user/user_followed_manga/user_followed_manga.dart
@@ -1,4 +1,4 @@
-import 'package:mangadex_library/src/models/common/data.dart';
+import 'package:mangadex_library/models/common/data.dart';
 
 class UserFollowedManga {
   late final List<Data> data;